prisma folder moving for the docker to find it on installation
